requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
ta>=0.11.0
python-dotenv>=1.0.0
langchain>=0.1.0
openai>=1.0.0
